fix: v0.100.0 — TC18 byte_message_info conformance (BREAKING) (#143)
* fix: v0.100.0 — TC18 byte_message_info conformance (BREAKING)

Rewrite acf.h/acf.c's byte_message_info header codec from the real
TC18 v0.5.1_RC Figure 7/Table 4 bit layout, replacing this project's
own invented layout (self-documented as such, never wire-compatible
with a real TC18 peer or with go-RCP/cpp-RCP): acf_msg_type (7 bits)
and acf_msg_length (9-bit QUADLET count over the whole message, not an
octet count of the payload alone) now share the first 16-bit word;
byte_bus_id is an 11-bit field split across two octets; the flag
octets are regrouped into the spec's own octet 2/4/6 layout. New
shared rcp_acf_pack_header()/rcp_acf_unpack_header() are the single
source of truth for this bit-packing, now used by acf.c itself and by
the five request-kind modules that build a raw ACF_GBB header by hand
to repurpose message_timestamp (each previously duplicated the wrong
bit-packing independently). e2e.c's adapt_acf_msg_length() and
scheduler.c's rcp_sched_split_frame_members() are updated to the new
field position.

Secondary fixes from the same header rework: mtv is now a single wire
bit (dropped the invented 3rd "uncertain" state); read_size_or_segment_num
widened to its real 12-bit range; a decoded byte_bus_id over 255 now
fails explicitly instead of truncating; op's encode-only NONE
convenience value now encodes identically to WRITE (there is only one
wire bit). RCP_ACF_MAX_PAYLOAD drops from 65535 to ~2028-2036; a
worst-case CAN XL new-payload request no longer fits a single
unfragmented ACF message (no fragmented request path exists yet,
tracked as a follow-up).

Hand-verified against the specification's own Figure 19 (ACF_ABB) and
Figure 20 (ACF_GBB) worked examples: both reproduce exactly
(acf_msg_length 0x05/0x07), pinned as golden-vector tests.

Closes TC18 §12.9.1.1's multi-request-per-frame gap: added
rcp_mock_server_dispatch_frame(), splitting a raw NTSCF/TSCF payload
into its constituent ACF messages and dispatching each to its own
addressed byte_bus_id.

Deferred: TC18's numbered wire error codes are still not populated
onto any endpoint's Error response (large per-endpoint effort, tracked
separately). coverage-report.json/qualify-report.json are cfusa-
generated artifacts whose fix belongs upstream in c-FuSa, not hand-
edited here. SHA-pinned ci.yml's ilammy/msvc-dev-cmd action. Added a
wire-interop hedge to README.md.

Full ctest suite 57/57 passing, clean under ASan/UBSan. Verified
locally against cfusa v0.5.50 (matching CI's pin): check/lint/cyber/
analyze/qualify/vuln all exit 0; trace 777/777 requirements traced and
tested (100%).


* fix: widen byte_bus_id before shifting to silence MSVC C4333

rcp_byte_bus_id_t is only 8 bits wide, so shifting it right by 8 in
rcp_acf_pack_header always yields 0 -- correct, but MSVC's /W4 /WX
flags the shift-by-operand-width as a data-loss warning (C4333) even
after usual arithmetic promotion to int, failing the Windows CI build.
Widen to uint16_t before shifting so the shift amount is provably
within the operand's width on every compiler.


* fix: make discovery's octet-width read_size/segment_num cast explicit

hdr.read_size_or_segment_num is now the wire header's real 12-bit
uint16_t field (this PR's own C-01/C-02 header rework), but
discovery's own read_size/segment_num was already, independently,
always one octet wide (RCP_DISCOVERY_GENERAL_SLICE_LEN's documented
invariant) -- the assignment was a same-width copy before this PR and
is still exactly as safe now, it just needs an explicit narrowing cast
for MSVC's /W4 to accept it the way GCC/Clang already silently did.


* fix: make ep_uart/ep_can's octet-width read_size/segment_num casts explicit

Same MSVC C4244 pattern as the discovery.c fix: hdr.read_size_or_segment_num
is now the wire header's real 12-bit uint16_t field, but ep_uart.c's
out_read_size and both files' local segment_num were already,
independently, uint8_t -- pre-existing narrower types this PR's header
rework didn't touch. The narrowing (and its truncation ceiling for a
read_size/segment_num above 255) already existed before this PR;
making the cast explicit doesn't change behavior, it just satisfies
MSVC's /W4 the way GCC/Clang already silently did.


---------

+++ src/discovery.c
@@ -149,8 +149,15 @@ rcp_discovery_errc_t rcp_discovery_decode_request(const uint8_t *b, size_t len,
     (void)payload;
     (void)payload_len;
 
+    /* hdr.read_size_or_segment_num is the wire header's general 12-bit
+     * field (TC18 Table 4), but discovery's own read_size is documented
+     * (see RCP_DISCOVERY_GENERAL_SLICE_LEN's comment in discovery.h) as
+     * always fitting one octet for this subsystem specifically -- the
+     * explicit cast makes that pre-existing, discovery-specific invariant
+     * visible to the compiler instead of relying on an implicit
+     * narrowing conversion. */
     out_req->requester       = requester;
-    out_req->read_size       = hdr.read_size_or_segment_num;
+    out_req->read_size       = (uint8_t)hdr.read_size_or_segment_num;
     out_req->transaction_num = hdr.transaction_num;
     return RCP_DISCOVERY_OK;
 }
@@ -332,8 +339,10 @@ rcp_discovery_errc_t rcp_discovery_decode_response_fragment(const uint8_t *b, si
     rc = decode_common(b, len, out_server_stream_id, &hdr, out_payload, out_payload_len);
     if (rc != RCP_DISCOVERY_OK) return rc;
 
+    /* Same discovery-specific octet-width invariant as
+     * rcp_discovery_decode_request()'s read_size cast above. */
     *out_ms          = (hdr.ms != 0u);
-    *out_segment_num = hdr.read_size_or_segment_num;
+    *out_segment_num = (uint8_t)hdr.read_size_or_segment_num;
     return RCP_DISCOVERY_OK;
 }
